Rework unit tests structure
Here we adjust the unit tests to the recent changes.
